Refactored package name from com.api.pojo to com.api.model
