+++ src/test/java/com/api/request/model/Problems.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.request.model;
 
 public record Problems (
