+++ src/test/java/com/api/request/model/UserCredentials.java
@@ -1,4 +1,4 @@
-package com.api.pojo;
+package com.api.request.model;
 
 public record UserCredentials(String username, String password) {
